Remove type hints from Hello QMI Notebook
Apologies for all the back and forth on this!

Per the discussion on the end of #5, the type hint changes I made
previously are 3.9 only—see [this
comment](https://github.com/rigetti/forest-tutorials/pull/5#issuecomment-726773787)
for details. As such, this PR removes type hints entirely from the Hello
QMI notebook.